fix: Cache names of players for the chat box
Previously, the chat would find the player's name from their ID each
time it wanted to display their message. This would throw an error in
case a player who sent messages in the chat decided to leave the game.

Now their names are cached so that they can be used even if the player
leaves the game.